src/pyproject.toml
[project]
name='com.castsoftware.uc.arg'
description="Assessment Report Generator (ARG)"
version='1.7.0' #prod version
authors= [
    {name="Nevin Kaplan",email="n.kaplan@castsoftware.com"},
    {name="DD Assessment Team",email="Team.ddassessments@castsoftware.com"}
]
readme="README.md"
dependencies = ['pandas', 'python-pptx==0.6.18', 'com.castsoftware.uc.python.common>=1.1.7', 'IPython', 'requests', 'Jinja2','setuptools']
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
requires-python = ">=3.6"
[project.urls]
Homepage = "https://github.com/CAST-Extend/com.castsoftware.uc.arg"
documentation="https://github.com/CAST-Extend/com.castsoftware.uc.arg/wiki"
